+++ viv-wallet-app/src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Illustration } from "./Illustration.vue";
+export { default as BalanceCard } from "./BalanceCard.vue";
+export { default as Banner } from "./Banner.vue";
+export { default as UserList } from "./UserList.vue";
